refactor: move event data fns to separate concern
